requirements.txt
fastapi>=0.115.13
pandas>=2.2.3
pydantic>=2.11.7
xgboost>=2.0.3
scikit-learn>=1.4.2
seaborn>=0.13.2
httpx>=0.27.0
uvicorn>=0.35.0
pytest>=8.4.1
pytest-cov>=4.1.0
google-cloud-storage>=3.2.0
python-dotenv>=1.1.1
locust>=2.37.14